Flatten lib dir
