--- packages/scrabble-solver/src/components/Board/lib/index.ts
@@ -1,3 +1,0 @@
-export { createGrid } from './createGrid';
-export { getBonusColor } from './getBonusColor';
-export { getPositionInGrid } from './getPositionInGrid';
